Make search bar font adjustable
TL;DR: UISearchBar sucks, implement your own!

The searchbar is not very well adjustable. So I wrote the whole thing
myself which looks beautiful now.

The left icon is gone since it was too much work to integrate it along
this change.

We also don't have a cancel button anymore but to be honest it is also
not really needed.