Use region code instead Locale in defaultCountry
Locale was kind of overkill as we only need the isoRegionCode to
determine the default country. It also enables persisting this code for
apps that e.g. use the country picker at login time. The persisted
code can then be used to re-initialize a phone number view controller
with an existing number when the user wants to edit it.

Apart from this i streamlined the wording to `isoCountryCode` when
referring to the two-letter [ISO 3166](https://en.wikipedia.org/wiki/List_of_ISO_3166_country_codes)
code of a country.